Better DI and preserve request object
Slim injects the dependency container every time a middleware is created from a class. This commit uses that instead of assuming that the `$next` callable is going to be an instance of `\Slim\App`.

Also, looks like there's no need to get the request object from the dependency container since the middleware is already receiving it as a parameter, otherwise any changes made to the PSR-7 request object are lost.

The above causes an issue for example when using `Slim-Csrf`because the request would no longer have "attributes" set by this middleware.

Hope it helps.

Regards!